requirements.txt
altgraph==0.17.3
bottle==0.12.25
bottle-websocket==0.2.9
cachetools==5.3.1
cffi==1.15.1
click==8.1.3
colorama==0.4.6
darkdetect==0.8.0
Eel==0.16.0
future==0.18.3
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.2
pefile==2023.2.7
Pillow==9.5.0
pycparser==2.21
pyinstaller==5.12.0
pyinstaller-hooks-contrib==2023.3
pyparsing==3.0.9
pywin32-ctypes==0.2.0
whichcraft==0.6.1
zope.event==4.6
zope.interface==6.0
